requirements-compatible.txt
# Core Flask Dependencies
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-socketio==5.3.6
flask-mail==0.9.1
flask-migrate==4.0.5
flask-cors==4.0.0

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1
eventlet==0.33.3

# Data Science & ML (Compatible versions)
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==1.7.6
torch>=2.2.0  # Compatible with Python 3.12
scipy==1.11.4

# Visualization & Reporting
plotly==5.17.0
reportlab==4.0.4

# Optional ML Features
shap==0.42.1
statsmodels==0.14.0
openai==0.28.1

# System & Utilities
apscheduler==3.10.4
prometheus-client==0.18.0
psutil==5.9.6
python-dotenv==1.0.0
gunicorn==21.2.0
python-fitbit==0.3.1